Manage the windows manually through our own windows controller: CallgrindOutputWindowController
Previously, a default controller was instantiated for us, and was loading the nib from windowNibName.
This offer little flexibility and give too many responsibilities to the CallgrindOutputDocument (it was
the first responder of the nib and supposed to be the target of the actions).

The window controller is now created and controlled by the application.

Reviewed by: Andreas Kling