Add a FileLoader class to read the file and decode the data
The FileLoader use a dispatch I/O channel to read the data of the file.

The read operations are done asynchronously in a low priority queue in order
to avoid slowing down the UI during the load.

The data read by the I/O queue is then split in lines for further processing. When
there is pending data before the next line break character, it is stored in a
temporary buffer until more data is read by the I/O channel.

If the document is closed before the file is completely read, the channel
is closed and the read is aborted.

Reviewed by: Andreas Kling