Add basic error management and show errors to the user
The file loader now has a callback to report errors to the user.
This mechanism is used to show an error dialog on the window.

Since canceling the load can cause an error, the call to cancel
cannot be done in dealloc (since the error handler retain the document)
so it is done on document::close instead.