Add a filter text input to filter the table by function name
The _functions attribute of the model now store the current
filtered function of the profile.

When the filter is reset, we take a shortcut and just
reload all the functions of the profile.

When the filter is not set, we start by finding what is the smallest
array we can filter. If the new filter is a more restrictive
than the current filter, we can use the current array. If not,
we use the full array of all functions.

The filter refuse the first responder due to a layout bug
in which the caret position is at a wrong position
when the view has the initial focus.